fix: db indexing & OrganizationsId refresh (#280)
* fix: db indexing & OrganizationsId refresh

* fix: format

* feat(organizer): Add organization creator as OWNER

This commit updates the organization creation process to automatically assign the `OWNER` role to the user who creates a new organization.

Previously, a new organization was created, and the user's profile was updated, but the user was not formally added as a member with a specific role within the new organization's member list.

This change ensures that upon successful organization creation:
- The creating user is added as a member with the `OWNER` role.
- Error handling is improved to provide more specific feedback if adding the member or updating the user's profile fails after the organization has been created.

* style: reformat the code with ktfmtFormat

---------

Co-authored-by: Zhixiang Dai <hnhnuh@126.com>