feat(search): Update user search to query `memberships` collection
This commit refactors the user search functionality to correctly exclude existing organization members.

Previously, the search logic queried a sub-collection (`organizations/{orgId}/members`) to identify users to exclude. This has been updated to query the root `memberships` collection using a `where` clause on the `orgId` field. This change aligns the search with the new data model where memberships are stored in a top-level collection.