test(auth): Move FakeAuthViewModel to separate file and add setSignedIn
* test(auth): Move FakeAuthViewModel to separate file

The `FakeAuthViewModel` class was previously defined inside `AuthComponentsTest.kt`. It has been removed from there and the existing separate file `FakeAuthViewModel.kt` is now being used, likely to improve reusability across tests.

* test(auth): Add setSignedIn method to FakeAuthViewModel

The `setSignedIn` helper method was added to the `FakeAuthViewModel` class to allow tests to easily simulate the signed-in state of the UI.